requirement.txt
requests
beautifulsoup4
tqdm
nltk
numpy
faiss-cpu
sentence-transformers
streamlit
